[BUG/#139] - 제휴 단체, 제휴 업체 목록 전체 조회 가능하도록 수정
