+++ app/src/main/java/com/ssu/assu/presentation/common/login/LoginActivity.kt
@@ -30,7 +30,6 @@ import kotlinx.coroutines.launch
 @AndroidEntryPoint
 class LoginActivity : BaseActivity<ActivityLoginBinding>(R.layout.activity_login) {
 
-    private val deviceTokenViewModel: DeviceTokenViewModel by viewModels()
     private val loginViewModel: LoginViewModel by viewModels()
     private var isAutoLoginChecked = false
 
@@ -61,9 +60,6 @@ class LoginActivity : BaseActivity<ActivityLoginBinding>(R.layout.activity_login
                 Toast.makeText(this, "이메일과 비밀번호를 입력해주세요.", Toast.LENGTH_SHORT).show()
                 return@setOnClickListener
             }
-
-            // ✅ 화면 종료 전에 토큰 등록까지 먼저 처리
-            fetchAndRegisterFcmToken()
 
             loginViewModel.commonLogin(email, password)
         }
@@ -104,34 +100,6 @@ class LoginActivity : BaseActivity<ActivityLoginBinding>(R.layout.activity_login
                 is LoginState.PendingApproval -> {
                     setLoginButtonEnabled(true)
                     Toast.makeText(this@LoginActivity, "승인 대기 중입니다: ${state.message}", Toast.LENGTH_SHORT).show()
-                }
-            }
-        }
-
-        // FCM 토큰 등록 상태 관찰
-        lifecycleScope.launch {
-            repeatOnLifecycle(Lifecycle.State.STARTED) {
-                deviceTokenViewModel.uiState.collect { state ->
-                    when (state) {
-                        is DeviceTokenViewModel.UiState.Idle -> Unit
-                        is DeviceTokenViewModel.UiState.Loading -> {
-                            // 필요하면 로딩 표시
-                            Log.d("FCM", "디바이스 토큰 등록 중…")
-                        }
-                        is DeviceTokenViewModel.UiState.Success -> {
-                            val tokenId = state.tokenId
-                            Log.i("FCM", "등록 성공: ${tokenId}")
-                        }
-                        is DeviceTokenViewModel.UiState.Fail -> {
-                            Log.e("FCM", "등록 실패: ${state.code} ${state.msg}")
-                            // FCM 토큰 등록 실패해도 앱을 종료하지 않음
-                            // 로그인은 성공했으므로 사용자가 계속 사용할 수 있도록 함
-                        }
-                        is DeviceTokenViewModel.UiState.Error -> {
-                            Log.e("FCM", "등록 오류: ${state.msg}")
-                            // FCM 토큰 등록 오류가 발생해도 앱을 종료하지 않음
-                        }
-                    }
                 }
             }
         }
@@ -192,26 +160,4 @@ class LoginActivity : BaseActivity<ActivityLoginBinding>(R.layout.activity_login
     private fun Int.dpToPx(context: Context): Int =
         (this * context.resources.displayMetrics.density).toInt()
 
-    //  서버 등록까지 한 번에
-    private fun fetchAndRegisterFcmToken() {
-        try {
-            FirebaseMessaging.getInstance().token.addOnCompleteListener { task ->
-                if (!task.isSuccessful) {
-                    Log.w("FCM", "토큰 가져오기 실패", task.exception)
-                    // FCM 토큰 가져오기 실패해도 앱을 종료하지 않음
-                    return@addOnCompleteListener
-                }
-                val token = task.result
-                if (token.isNullOrEmpty()) {
-                    Log.w("FCM", "FCM 토큰이 비어있음")
-                    return@addOnCompleteListener
-                }
-                Log.d("FCM", "FCM 토큰: $token")
-                deviceTokenViewModel.register(token)
-            }
-        } catch (e: Exception) {
-            Log.e("FCM", "FCM 토큰 등록 중 예외 발생", e)
-            // 예외가 발생해도 앱을 종료하지 않음
-        }
-    }
 }
